+++ ClHcaSharp/Channel.cs
@@ -5,14 +5,14 @@ namespace ClHcaSharp
     internal class Channel
     {
         public ChannelType Type { get; set; }
-        public uint CodedCount { get; set; }
+        public int CodedCount { get; set; }
 
         public byte[] Intensity { get; } = new byte[SubframesPerFrame];
         public byte[] ScaleFactors { get; } = new byte[SamplesPerSubframe];
         public byte[] Resolution { get; } = new byte[SamplesPerSubframe];
         public byte[] Noises { get; } = new byte[SamplesPerSubframe];
-        public uint NoiseCount { get; set; }
-        public uint ValidCount { get; set; }
+        public int NoiseCount { get; set; }
+        public int ValidCount { get; set; }
 
         public float[] Gain { get; } = new float[SamplesPerSubframe];
         public float[] Spectra { get; } = new float[SamplesPerSubframe];
@@ -22,6 +22,6 @@ namespace ClHcaSharp
 
         public float[][] Wave { get; } =
             JaggedArray.CreateJaggedArray<float[][]>(
-                (int)SubframesPerFrame, (int)SamplesPerSubframe);
+                SubframesPerFrame, SamplesPerSubframe);
     }
 }
